+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.9.4
+* Closing & disposing fixes. Fixed android camera facing.
+* Fix: Android - Pass audioSettings.muted to RtmpStream.
+
+## 0.9.3
+* Bump up HaishinKit.kt to 0.10.4
+* Improved compatibility some services.
+
+## 0.9.2
+* Fix setAudioSettings were used instead of setVideoSettings.
+
 ## 0.9.1
 * Support rotate with a device orientation.
 * Support dispose platform interface.